curl: Support CURLOPT_LOCALADDR to bind to local address.
This allows a user to bind to both an interface (with CURLOPT_INTERFACE)
and a local IP address.  In doing so, it allows the user to work-around
some serious performance issues on machines with lots of virtual interfaces
because curl no longer has to scan all interfaces each time it makes
a connection.

Signed-off-by: Ben Greear <greearb@candelatech.com>