connect:  Fix case where addr but not dev was configured.
And add a much more verbose failure message so folks can
better figure out why it died.  In my case, it was because
I was resolving as 'whatever' but binding to IPv4 addrs.
I was getting an AF_INET6 address from the resolver
until I forced resolving to be ipv4.

Signed-off-by: Ben Greear <greearb@candelatech.com>