tool_getenv: Add new functions for Windows Unicode builds
- For Windows Unicode builds add tool_getenv_utf8 to get Unicode UTF-8
  encoded environment variable values.

- Add tool_getenv_local to getenv in the current locale encoding. This
  is the equivalent of curl_getenv which is now banned in favor of
  tool_getenv.

- Map tool_getenv macro to tool_getenv_utf8 for Windows Unicode and
  tool_getenv_local otherwise.

- Similar to above, split homedir into homedir_utf8, homedir_local
  and a homedir macro which maps to one of the first two.

Background:

Windows does not support UTF-8 locale (or, not really). Because of that
our Windows Unicode builds continue to use the current locale, but
expect Unicode UTF-8 encoded paths for internal use. In other words file
operations by curl or libcurl to open / access / stat a file are
expected to have a Unicode path.

Complicating this is that dependencies, which may or may not be Unicode
builds, may or may not expect UTF-8 encoded Unicode for char * string
pathnames. For example, libcurl can use libssh or libssh2 as an SSH
library and pathnames need to be passed in the local encoding AFAICS.

Prior to this change the curl tool made a best effort to convert
pathnames from the command line to Unicode UTF-8 encoding, but did not
do so for those pathnames retrieved by environment variables via
curl_getenv, which is in the local encoding. This worked without much
incident but was incorrect.

Essentially, for Windows Unicode builds we need access to both local
encoded and UTF-8 encoded pathnames from the environment. Therefore,
curl_getenv is not sufficient. In order to make this obvious in the
code I've banned curl_getenv (via checksrc) in favor of new
tool_getenv_local, tool_getenv_utf8 and tool_getenv and made similar
changes to homedir, as described.

The end result is something like CURLOPT_SSH_KNOWNHOSTS which is passed
to the SSH library and always needs the local encoding (in other words,
even if it's a Windows Unicode build) can now be made by calling
homedir_local (which calls tool_getenv_local). But, if it's a path that
is used internally only then just homedir can be used (which calls
tool_getenv_utf8 for Windows Unicode, and tool_getenv_local otherwise).

Also of note is 765e060, which removed local encoding fallbacks and
is another reason to make these changes.

Fixes https://github.com/curl/curl/pull/7252
Closes https://github.com/curl/curl/pull/7281